finish code coverage

+++ src/Type/Node.php
@@ -8,6 +8,7 @@ use Syndesi\CypherDataStructures\Contract\NodeInterface;
 use Syndesi\CypherDataStructures\Contract\NodeLabelInterface;
 use Syndesi\CypherDataStructures\Contract\NodeLabelStorageInterface;
 use Syndesi\CypherDataStructures\Contract\RelationInterface;
+use Syndesi\CypherDataStructures\Contract\WeakRelationInterface;
 use Syndesi\CypherDataStructures\Contract\WeakRelationStorageInterface;
 use Syndesi\CypherDataStructures\Exception\InvalidArgumentException;
 use Syndesi\CypherDataStructures\Helper\ToCypherHelper;
@@ -79,31 +80,60 @@ class Node implements NodeInterface
     /**
      * @throws InvalidArgumentException
      */
-    public function addRelation(RelationInterface $relation): self
+    public function addRelation(WeakRelationInterface|RelationInterface $relation): self
     {
+        $weakRelation = $relation;
+        if ($weakRelation instanceof RelationInterface) {
+            $weakRelation = WeakRelation::create($weakRelation);
+        }
+        if (null === $weakRelation->get()) {
+            throw InvalidArgumentException::createForAlreadyNullReference(WeakRelationInterface::class);
+        }
+        /** @psalm-suppress PossiblyNullReference */
+        if (null === $weakRelation->get()->getStartNode()) {
+            throw InvalidArgumentException::createForTypeMismatch(NodeInterface::class, 'null');
+        }
+        /** @psalm-suppress PossiblyNullReference */
+        if (null === $weakRelation->get()->getEndNode()) {
+            throw InvalidArgumentException::createForTypeMismatch(NodeInterface::class, 'null');
+        }
         $ownIdentifyingString = ToCypherHelper::nodeToIdentifyingCypherString($this);
-        if (ToCypherHelper::nodeToIdentifyingCypherString($relation->getStartNode()) !== $ownIdentifyingString &&
-            ToCypherHelper::nodeToIdentifyingCypherString($relation->getEndNode()) !== $ownIdentifyingString
+        /** @psalm-suppress PossiblyNullReference */
+        if (ToCypherHelper::nodeToIdentifyingCypherString($weakRelation->get()->getStartNode()) !== $ownIdentifyingString &&
+            ToCypherHelper::nodeToIdentifyingCypherString($weakRelation->get()->getEndNode()) !== $ownIdentifyingString
         ) {
             throw new InvalidArgumentException("Adding a relation to a node requires that either the start node or the end node must be the same as the node itself.");
         }
 
-        $this->weakRelationStorage->attach($relation);
+        $this->weakRelationStorage->attach($weakRelation);
 
         return $this;
     }
 
+    /**
+     * @throws InvalidArgumentException
+     */
     public function addRelations(WeakRelationStorageInterface $weakRelationStorage): self
     {
         foreach ($weakRelationStorage as $key) {
-            $this->weakRelationStorage->attach($key);
+            $this->addRelation($key);
         }
 
         return $this;
     }
 
-    public function hasRelation(RelationInterface $relation): bool
+    /**
+     * @throws InvalidArgumentException
+     */
+    public function hasRelation(WeakRelationInterface|RelationInterface $relation): bool
     {
+        if ($relation instanceof RelationInterface) {
+            $relation = WeakRelation::create($relation);
+        }
+        if (null === $relation->get()) {
+            throw InvalidArgumentException::createForAlreadyNullReference(WeakRelationInterface::class);
+        }
+
         return $this->weakRelationStorage->contains($relation);
     }
 
@@ -112,8 +142,17 @@ class Node implements NodeInterface
         return $this->weakRelationStorage;
     }
 
-    public function removeRelation(RelationInterface $relation): self
+    /**
+     * @throws InvalidArgumentException
+     */
+    public function removeRelation(WeakRelationInterface|RelationInterface $relation): self
     {
+        if ($relation instanceof RelationInterface) {
+            $relation = WeakRelation::create($relation);
+        }
+        if (null === $relation->get()) {
+            throw InvalidArgumentException::createForAlreadyNullReference(WeakRelationInterface::class);
+        }
         $this->weakRelationStorage->detach($relation);
 
         return $this;
